test server for better visualization of the ai
to better display the process of the ai a webserver has been programmed for the user. with that webserver the user is able to see the drawing process and in addition can edit images direct in the browser so there is no need for saving the images which have been drawn. in addition the images get auto documented